# openapi.json
{"openapi":"3.1.0","info":{"title":"Urja Meter API","description":"Clean API wrapper over the legacy Urja Meter Ops portal.","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Health Health Get"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Root  Get"}}}}}}},"/api/v1/meters":{"get":{"summary":"Search Meters","operationId":"search_meters_api_v1_meters_get","parameters":[{"name":"search","in":"query","required":false,"schema":{"type":"string","default":"","title":"Search"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeterList"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/meters/{meter_id}":{"get":{"summary":"Get Meter","operationId":"get_meter_api_v1_meters__meter_id__get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","minLength":1,"maxLength":50,"title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Meter"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/meters/{meter_id}/consumption":{"get":{"summary":"Get Consumption","operationId":"get_consumption_api_v1_meters__meter_id__consumption_get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","minLength":1,"maxLength":50,"title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsumptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/meters/{meter_id}/location":{"get":{"summary":"Get Location","operationId":"get_location_api_v1_meters__meter_id__location_get","parameters":[{"name":"meter_id","in":"path","required":true,"schema":{"type":"string","minLength":1,"maxLength":50,"title":"Meter Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ConsumptionResponse":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"readings":{"items":{"$ref":"#/components/schemas/EnergyReading"},"type":"array","title":"Readings"}},"type":"object","required":["meter_id","readings"],"title":"ConsumptionResponse"},"EnergyReading":{"properties":{"timestamp":{"type":"string","title":"Timestamp"},"kwh":{"type":"number","title":"Kwh"},"kvah":{"type":"number","title":"Kvah"},"voltage_r":{"type":"number","title":"Voltage R"}},"type":"object","required":["timestamp","kwh","kvah","voltage_r"],"title":"EnergyReading"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LocationResponse":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"latitude":{"type":"number","title":"Latitude"},"longitude":{"type":"number","title":"Longitude"}},"type":"object","required":["meter_id","latitude","longitude"],"title":"LocationResponse"},"Meter":{"properties":{"meter_id":{"type":"string","title":"Meter Id"},"serial_number":{"type":"string","title":"Serial Number"},"make":{"type":"string","title":"Make"},"phase_type":{"type":"string","title":"Phase Type"},"installation_status":{"type":"string","title":"Installation Status"},"distribution_transformer":{"type":"string","title":"Distribution Transformer"}},"type":"object","required":["meter_id","serial_number","make","phase_type","installation_status","distribution_transformer"],"title":"Meter"},"MeterList":{"properties":{"data":{"items":{"$ref":"#/components/schemas/Meter"},"type":"array","title":"Data"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"page_size":{"type":"integer","title":"Page Size"}},"type":"object","required":["data","total","page","page_size"],"title":"MeterList"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}